nexus-api/requirements.txt
fastapi>=0.115
uvicorn>=0.34
pydantic>=2.11
sentence-transformers>=3.3
numpy>=2.2
pyyaml>=6.0
anthropic>=0.40
